+++ STM8/peripherals/tim2.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 void tim2_init();
-void tim2_clockrun(uint16_t clocks);
+void tim2_clockrun(int32_t clocks);
 void tim2_write(uint32_t addr, uint8_t val);
 uint8_t tim2_read(uint32_t addr);
